Roughly impemented UI with rudimentary file import ('upload'), you can drag and drop your .wav files in the dropzone which then populates a list of 'samples' in the GUI. you can re-order, or remove the samples through the GUI and then 'export' the list in order from top to bottom with the "export kit" button.
yet to implement:
preset generation from an .xml template
zipping up preset with the .wav samples
playback of the samples through a play button
waveform generation of the sample with a playhead
and 'advanced' mode for custom filepaths instead of auto-generated ones.
a 'semi-advanced' mode for generating filepaths based on sample categories set by the user. i.e (one-shot > kick drum) which would then create those subfolders and place the sample within it. this should allow 're-usability' of the samples between presets while not dumping everything in one folder.